Refactor helpers into dedicated modules
