+++ operators/tracking/solver.py
@@ -1294,141 +1294,6 @@ class CLIP_OT_track_sequence(bpy.types.Operator):
         return {'FINISHED'}
 
 
-def has_active_marker(tracks, frame):
-    for t in tracks:
-        m = t.markers.find_frame(frame)
-        if m and not m.mute and m.co.length_squared != 0.0:
-            return True
-    return False
-
-
-def get_undertracked_markers(clip, min_frames=10):
-    undertracked = []
-
-    clean_pending_tracks(clip)
-
-    for track in clip.tracking.tracks:
-        if not (track.name.startswith(PREFIX_TRACK) or track in PENDING_RENAME):
-            continue
-
-        tracked_frames = [
-            m for m in track.markers
-            if not m.mute and m.co.length_squared != 0.0
-        ]
-
-        if len(tracked_frames) < min_frames:
-            undertracked.append((track.name, len(tracked_frames)))
-
-    return undertracked
-
-
-def select_tracks_by_names(clip, name_list):
-    for track in clip.tracking.tracks:
-        track.select = track.name in name_list
-
-
-def select_tracks_by_prefix(clip, prefix):
-    """Select all tracks whose names start with the given prefix."""
-    for track in clip.tracking.tracks:
-        track.select = track.name.startswith(prefix)
-
-
-def ensure_valid_selection(clip, frame):
-    """Validate selected tracks for the given frame."""
-    valid = False
-    for track in clip.tracking.tracks:
-        if track.select:
-            marker = track.markers.find_frame(frame, exact=True)
-            if marker is None or marker.mute:
-                track.select = False
-            else:
-                valid = True
-    return valid
-
-
-def track_markers_range(scene, start, end, current, backwards):
-    """Run clip.track_markers for ``start`` to ``end``."""
-    scene.frame_start = start
-    scene.frame_end = end
-    if not backwards:
-        scene.frame_current = current
-        update_frame_display()
-    bpy.ops.clip.track_markers(backwards=backwards, sequence=True)
-
-
-def jump_to_first_frame_with_few_active_markers(min_required=5):
-    scene = bpy.context.scene
-    clip = bpy.context.space_data.clip
-
-    for frame in range(scene.frame_start, scene.frame_end + 1):
-        count = 0
-        for track in clip.tracking.tracks:
-            if track.name.startswith(PREFIX_GOOD):
-                marker = track.markers.find_frame(frame)
-                if marker and not marker.mute and marker.co.length_squared != 0.0:
-                    count += 1
-
-        if count < min_required:
-            jump_to_frame_with_few_markers(
-                clip,
-                min_marker_count=min_required,
-                start_frame=frame,
-                end_frame=frame,
-            )
-            update_frame_display()
-            _update_nf_and_motion_model(frame, clip)
-            return frame
-
-    return None
-
-
-def find_low_marker_frame(clip, threshold):
-    """Return the first frame with fewer markers than ``threshold``."""
-    scene = bpy.context.scene
-    for frame in range(scene.frame_start, scene.frame_end + 1):
-        count = 0
-        for track in clip.tracking.tracks:
-            marker = track.markers.find_frame(frame)
-            if marker and not marker.mute and marker.co.length_squared != 0.0:
-                count += 1
-        if count < threshold:
-            return frame, count
-    return None, None
-
-
-
-
-def _update_nf_and_motion_model(frame, clip):
-    """Maintain NF list and adjust motion model and pattern size.
-
-    Wenn die Pattern Size 100 erreicht, wird stattdessen der Wert aus
-    ``Scene.marker_frame`` um 10 % erh\u00f6ht (maximal das Doppelte des
-    Ausgangswerts). Sinkt die Pattern Size wieder unter 100, verkleinert sich
-    ``marker_frame`` schrittweise um 10 %, bis der Startwert erreicht ist.
-    """
-
-    global NF
-    settings = clip.tracking.settings
-    scene = bpy.context.scene
-    min_size, max_size = pattern_limits(clip)
-    if frame in NF:
-        cycle_motion_model(settings, clip, reset_size=False)
-        if settings.default_pattern_size < max_size:
-            settings.default_pattern_size = min(
-                int(settings.default_pattern_size * 1.1),
-                max_size,
-            )
-        else:
-            max_mf = DEFAULT_MARKER_FRAME * 2
-            scene.marker_frame = min(int(scene.marker_frame * 1.1), max_mf)
-    else:
-        NF.append(frame)
-        settings.default_motion_model = DEFAULT_MOTION_MODEL
-        settings.default_pattern_size = int(settings.default_pattern_size * 0.9)
-        if settings.default_pattern_size < max_size and scene.marker_frame > DEFAULT_MARKER_FRAME:
-            scene.marker_frame = max(int(scene.marker_frame * 0.9), DEFAULT_MARKER_FRAME)
-    settings.default_pattern_size = clamp_pattern_size(settings.default_pattern_size, clip)
-    settings.default_search_size = settings.default_pattern_size * 2
 
 def _Test_detect(self, context, use_defaults=True):
     """Run the Test detect cycle optionally using default settings."""
@@ -1808,68 +1673,6 @@ class CLIP_OT_good_marker_position(bpy.types.Operator):
 
 
 
-
-def enable_proxy():
-    """Enable proxies if possible."""
-    if bpy.ops.clip.proxy_on.poll():
-        bpy.ops.clip.proxy_on()
-
-
-def detect_features_once(context=None, clip=None, threshold=None):
-    """Run feature detection if available."""
-    if context is None:
-        context = bpy.context
-    if clip is None:
-        clip = getattr(context.space_data, "clip", None)
-    if clip is None:
-        print("\u26a0\ufe0f Kein aktiver Movie Clip gefunden.")
-        return
-    if threshold is None:
-        threshold = context.scene.tracker_threshold
-
-    if bpy.ops.clip.detect_features.poll():
-        width, _ = clip.size
-        if width == 0:
-            print("\u26a0\ufe0f Clipgr\u00f6\u00dfe ist 0 - Detect abgebrochen")
-            return
-
-        margin_base, min_distance_base = calculate_base_values(width)
-        print(
-            f"[BASE DEBUG] width={width}, margin_base={margin_base}, min_distance_base={min_distance_base}"
-        )
-
-        detection_threshold, margin, min_distance = compute_detection_params(
-            threshold, margin_base, min_distance_base
-        )
-
-        print(
-            f"[Detect Features] threshold={threshold:.8f}, "
-            f"margin={margin}, min_distance={min_distance}"
-        )
-        if bpy.ops.clip.proxy_off.poll():
-            bpy.ops.clip.proxy_off()
-        bpy.ops.clip.detect_features(
-            threshold=threshold,
-            margin=margin,
-            min_distance=min_distance,
-        )
-        print(
-            f"[Detect Features] finished with threshold {threshold:.8f}"
-        )
-
-
-def detect_features_main(context, clip, threshold):
-    """Run feature detection with the aggressive marker target."""
-    target = marker_target_aggressive(context.scene.marker_frame)
-    detect_features_once(context=context, clip=clip, threshold=threshold)
-    return target
-
-
-def detect_features_test(context, clip, threshold):
-    """Run feature detection with the conservative marker target."""
-    target = marker_target_conservative(context.scene.marker_frame)
-    detect_features_once(context=context, clip=clip, threshold=threshold)
-    return target
